# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /measurement-records:
    get:
      operationId: listMeasurementRecords
      tags: [orders]
      summary: List measurement sheets and retained versions
      responses:
        "200":
          description: Measurement sheet records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementRecordList"
    post:
      operationId: createMeasurementRecord
      tags: [orders]
      summary: Create a measurement sheet record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementRecordInput"
      responses:
        "201":
          description: Measurement record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementRecord"
  /measurement-records/lookup:
    get:
      operationId: searchMeasurementRecords
      tags: [orders]
      summary: Search measurement sheets for quotation linking
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      responses:
        "200":
          description: Matching measurement records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementRecordLookupList"
  /measurement-records/{recordId}:
    patch:
      operationId: updateMeasurementRecord
      tags: [orders]
      summary: Update measurement record details or assign it to an order
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeasurementRecordUpdate"
      responses:
        "200":
          description: Updated measurement record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementRecord"
        "404":
          description: Measurement record or order not found
  /measurement-records/{recordId}/versions/{filename}:
    post:
      operationId: uploadMeasurementVersion
      tags: [orders]
      summary: Add a retained measurement sheet version
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 180
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: New measurement version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasurementVersion"
        "400":
          description: Invalid measurement sheet
        "404":
          description: Measurement record not found
  /measurement-records/{recordId}/versions/{versionId}/content:
    get:
      operationId: downloadMeasurementVersion
      tags: [orders]
      summary: Download a retained measurement sheet version
      parameters:
        - name: recordId
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Measurement sheet file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Version not found
...
components:
  schemas:
...
    MeasurementVersion:
      type: object
      properties:
        id:
          type: string
        versionNumber:
          type: integer
        filename:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        uploadedBy:
          type: string
        uploadedByName:
          type: string
        uploadedAt:
          type: string
          format: date-time
      required: [id, versionNumber, filename, contentType, sizeBytes, uploadedBy, uploadedByName, uploadedAt]
    MeasurementRecord:
      type: object
      properties:
        id:
          type: string
        clientName:
          type: string
        location:
          type: ["string", "null"]
        orderRecordId:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
        versions:
          type: array
          items:
            $ref: "#/components/schemas/MeasurementVersion"
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, clientName, location, orderRecordId, orderId, versions, createdBy, createdAt, updatedAt]
    MeasurementRecordList:
      type: array
      items:
        $ref: "#/components/schemas/MeasurementRecord"
    MeasurementRecordLookup:
      type: object
      properties:
        id:
          type: string
        clientName:
          type: string
        location:
          type: ["string", "null"]
        orderId:
          type: ["string", "null"]
      required: [id, clientName, location, orderId]
    MeasurementRecordLookupList:
      type: array
      items:
        $ref: "#/components/schemas/MeasurementRecordLookup"
    MeasurementRecordInput:
      type: object
      properties:
        clientName:
          type: string
          minLength: 1
          maxLength: 160
        location:
          type: ["string", "null"]
          maxLength: 160
        orderRecordId:
          type: ["string", "null"]
      required: [clientName, location, orderRecordId]
    MeasurementRecordUpdate:
      type: object
      properties:
        clientName:
          type: string
          minLength: 1
          maxLength: 160
        location:
          type: ["string", "null"]
          maxLength: 160
        orderRecordId:
          type: ["string", "null"]
        linkQuotationSubmissionId:
          type: string